+++ src/main/java/com/github/lipinskipawel/board/ai/BoardEvaluator.java
@@ -1,9 +1,9 @@
 package com.github.lipinskipawel.board.ai;
 
-import com.github.lipinskipawel.board.engine.BoardInterface;
+import com.github.lipinskipawel.board.engine.Board;
 
 public interface BoardEvaluator {
 
-    double evaluate(BoardInterface board);
+    double evaluate(Board board);
 
 }