+++ src/main/java/com/github/lipinskipawel/board/engine/Boards.java
@@ -4,7 +4,7 @@ package com.github.lipinskipawel.board.engine;
 public class Boards {
 
 
-    public static BoardInterface immutableBoard() {
+    public static Board immutableBoard() {
         return new ImmutableBoard();
     }