+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - The MiniMax class is no longer a part of the API
 - The DummyBoardEvaluator is no longer a part of the API
 
+### Rename
+- BoardInterface into Board
+
 ### Added
 - Default method providing builder for MoveStrategy